+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [Unreleased]
+
+### Added
+- Create database indexes (from @justjasongreen)
+
+
 ## [1.0.0a1] - 2016-07-22
 
 ### Added
@@ -32,5 +38,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Set up project (from @justjasongreen)
 
 
+[Unreleased]: https://github.com/justjasongreen/racing_data/compare/1.0.0a1...HEAD
 [1.0.0a1]: https://github.com/justjasongreen/racing_data/compare/1.0.0a0...1.0.0a1
 [1.0.0a0]: https://github.com/justjasongreen/racing_data/tree/1.0.0a0